+++ Assets/Scripts/General/TextureManager.cs
@@ -300,7 +300,7 @@ namespace General
             GetTextureFromRender(input, mapType, out _);
         }
 
-        private void GetTextureFromRender(RenderTexture input, out Texture2D outTexture)
+        public void GetTextureFromRender(RenderTexture input, out Texture2D outTexture)
         {
             GetTextureFromRender(input, ProgramEnums.MapType.None, out outTexture);
         }